New functions , GUI, optimizations
In this version there are optimizations on the code + a few functions and operations added.
The code has been optimized for better reading and understanding of the code and comments were added were neccessary to make it even easier.
Also there is a starting point on working on developing a GUI.